[bugfix] Refactoring pull request #89 after code-review.
This commit moves the logic added into the RepositoryService
to the RepositoryController.

*	modified:   grails-app/controllers/docker/registry/web/RepositoryController.groovy
- Adding the logic from the service to the controller from the previous pull request commit.

*	modified:   grails-app/services/docker/registry/web/RepositoryService.groovy
- Reverting the service class to the previous version.
